make index type + dataID type in update code explicit
The dataID is only a 32 bit unsigned integer in the MySQL
database, so a larger type is not necessary.

+++ src/update/UpdateTo_0.5.5.cpp
@@ -131,7 +131,7 @@ bool UpdateTo055::updateData(const ConnectionInformation& ci)
              continue;
       } // switch
       Weather w;
-      const unsigned int dataId = row.getInt64(0);
+      const uint32_t dataId = static_cast<uint32_t>(row.getInt64(0));
       if (!api->parseCurrentWeather(row[2].c_str(), w))
       {
         std::cerr << "Error: Could not parse JSON data for data ID " << dataId
